+++ packages/main/base.mjs
@@ -1,5 +1,5 @@
 import { eleX, createXEle } from "./util.mjs";
-import { render, convert } from "./render.mjs";
+import { render, convert } from "./render/render.mjs";
 import Xhear from "./main.mjs";
 import stanz from "../stanz/src/base.mjs";
 import { getType } from "../stanz/src/public.mjs";
