+++ packages/main/render/fill.mjs
@@ -1,6 +1,6 @@
-import Stanz from "../stanz/src/main.mjs";
-import { hyphenToUpperCase, isArrayEqual } from "./public.mjs";
-import { createXEle } from "./util.mjs";
+import Stanz from "../../stanz/src/main.mjs";
+import { hyphenToUpperCase, isArrayEqual } from "../public.mjs";
+import { createXEle } from "../util.mjs";
 import { render } from "./render.mjs";
 
 export default {
