+++ packages/main/render/render.mjs
@@ -1,5 +1,5 @@
-import { isFunction, hyphenToUpperCase } from "./public.mjs";
-import { eleX } from "./util.mjs";
+import { isFunction, hyphenToUpperCase } from "../public.mjs";
+import { eleX } from "../util.mjs";
 
 const searchEle = (el, expr) => Array.from(el.querySelectorAll(expr));
